backend/requirements.txt
# Web框架
Flask==2.3.3
Flask-RESTful==0.3.10
Flask-RESTX==1.3.0
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6

# 数据库
SQLAlchemy==2.0.23
psycopg2-binary==2.9.7
alembic==1.12.1

# 异步任务
celery==5.3.4
redis==5.0.1

# FTP客户端
paramiko==3.3.1

# 数据验证
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0

# 密码加密
bcrypt==4.0.1
cryptography==41.0.7

# 工具库
python-dotenv==1.0.0
click==8.1.7
Werkzeug==2.3.7

# 监控和日志
python-json-logger==2.0.7

# 开发工具
pytest==7.4.3
pytest-flask==1.3.0
black==23.9.1
flake8==6.1.0

# 生产环境
gunicorn==21.2.0
eventlet==0.33.3
